GCP/gcp_umi_function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage >=1.25.0
google-cloud-pubsub >=1.1.0
#google-cloud-error-reporting >= 0.33.0